refactor: move the applicationsettings to the settings package
Move ApplicationSettings to go along with the Theme enum class, any other settings that the
application can interact with should be pulled from here
