+++ app-search/src/main/java/com/chesire/nekome/app/search/results/ResultsViewModel.kt
@@ -3,10 +3,10 @@ package com.chesire.nekome.app.search.results
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import com.chesire.nekome.core.ApplicationSettings
 import com.chesire.nekome.core.AuthCaster
 import com.chesire.nekome.core.flags.SeriesType
 import com.chesire.nekome.core.models.SeriesModel
+import com.chesire.nekome.core.settings.ApplicationSettings
 import com.chesire.nekome.series.SeriesRepository
 import com.chesire.nekome.server.Resource
 import kotlinx.coroutines.launch
