+++ core/src/main/java/com/chesire/nekome/core/settings/ApplicationSettings.kt
@@ -1,10 +1,10 @@
-package com.chesire.nekome.core
+package com.chesire.nekome.core.settings
 
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.core.content.edit
+import com.chesire.nekome.core.R
 import com.chesire.nekome.core.flags.UserSeriesStatus
-import com.chesire.nekome.core.settings.Theme
 import javax.inject.Inject
 
 /**
